requirements.txt
binance-connector==1.18.0